developed new interface
